Add sysroot to ease usage
Having sysroot included eases the process of cross-compiling Linux
applications. In order to make it compatible with both linux and windows
host systems, the symbolic links have all been replaced with hard copies
of the files they point to. This adds ~40MB to the checked out size, but
removes the added complexity of having to distribute the sysroot separately
and including decompression instructions.